feat: move the verifyLogin method to the auth controller
